# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NSG MOA API Documentation",
        "description": "Nasarawa State Ministry of Agriculture M&E System API documentation",
        "contact": {
            "email": "support@nsgmoa.gov.ng"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/farmers": {
            "get": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Get list of farmers",
                "operationId": "8044bbebf065e37fc63db6926112a802",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of farmers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Farmer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Create a new farmer",
                "operationId": "a9db659906f41717b6dd13ee640fb6a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFarmerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Farmer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farmer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/farmers/{id}": {
            "get": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Get specific farmer",
                "operationId": "94fc9dd79745ee6ad8db5016f0887d73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Farmer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farmer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Update farmer",
                "operationId": "0c3b45b2666751f6adb3840f8801568c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFarmerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Farmer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farmer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Delete farmer",
                "operationId": "e543d24a945dbfd7e9f738b3773bc567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Farmer deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/farmers/nearby/{lat}/{lng}/{distance}": {
            "get": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Find farmers within distance",
                "operationId": "5a5d45c2ffbbcb483444c4b844972416",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "distance",
                        "in": "path",
                        "description": "Distance in meters",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of nearby farmers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Farmer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/farmers/in-area": {
            "post": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Find farmers in area",
                "operationId": "c44bdc159b6216bd4a9d63f43266c087",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "area": {
                                        "description": "GeoJSON polygon",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of farmers in area",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Farmer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreFarmerRequest": {
                "description": "Farmer creation request",
                "required": [
                    "registration_number",
                    "full_name",
                    "phone_number",
                    "location_coordinates",
                    "farm_size",
                    "registration_date",
                    "status"
                ],
                "properties": {
                    "registration_number": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "full_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "location_coordinates": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "enum": [
                                    "Point"
                                ]
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "number"
                                },
                                "maxItems": 2,
                                "minItems": 2
                            }
                        },
                        "type": "object"
                    },
                    "farm_size": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "pending"
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateFarmerRequest": {
                "description": "Farmer update request",
                "required": [
                    "registration_number",
                    "full_name",
                    "phone_number",
                    "location_coordinates",
                    "farm_size",
                    "registration_date",
                    "status"
                ],
                "properties": {
                    "registration_number": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "full_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "location_coordinates": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "enum": [
                                    "Point"
                                ]
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "number"
                                },
                                "maxItems": 2,
                                "minItems": 2
                            }
                        },
                        "type": "object"
                    },
                    "farm_size": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "pending"
                        ]
                    }
                },
                "type": "object"
            },
            "Farmer": {
                "description": "Farmer model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "registration_number": {
                        "type": "string"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "location_geojson": {
                        "type": "object"
                    },
                    "farm_size": {
                        "type": "number",
                        "format": "float"
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "primary_crop": {
                        "type": "string"
                    },
                    "years_farming": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}